Move mapper to repo or controller.
