+++ apps/server/src/modules/account/repo/mapper/account-entity-to-do.mapper.ts
@@ -1,6 +1,6 @@
 import { AccountEntity } from '@shared/domain/entity';
 import { Counted } from '@shared/domain/types';
-import { Account } from '../domain/account';
+import { Account } from '../../domain/account';
 
 export class AccountEntityToDoMapper {
 	static mapToDto(account: AccountEntity): Account {
