+++ apps/server/src/modules/account/repo/mapper/account-idm-to-do.mapper.abstract.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { IdmAccount } from '@shared/domain/interface';
-import { Account } from '../domain/account';
+import { Account } from '../../domain/account';
 
 @Injectable()
 export abstract class AccountIdmToDoMapper {
