+++ apps/server/src/modules/account/repo/mapper/account-idm-to-do.mapper.db.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { IdmAccount } from '@shared/domain/interface';
-import { Account } from '../domain';
+import { Account } from '../../domain';
 import { AccountIdmToDoMapper } from './account-idm-to-do.mapper.abstract';
 import { AccountIdmToDoMapperDb } from './account-idm-to-do.mapper.db';
