+++ apps/server/src/modules/account/repo/mapper/index.ts
@@ -2,4 +2,3 @@ export * from './account-idm-to-do.mapper.abstract';
 export * from './account-idm-to-do.mapper.idm';
 export * from './account-idm-to-do.mapper.db';
 export * from './account-entity-to-do.mapper';
-export * from './account-response.mapper';
