+++ apps/server/src/modules/account/review-comments.md
@@ -1,6 +1,5 @@
 # Review Comments 14.7.23
 
-- move mapper into repo folder
 - write an md file or flow diagram describing how things work
 - in what layer do the services belong?
