+++ apps/server/src/modules/provisioning/strategy/oidc/service/oidc-provisioning.service.spec.ts
@@ -1,7 +1,7 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { AccountService } from '@modules/account/services/account.service';
-import { AccountSaveDto } from '@src/modules/account/domain';
+import { Account } from '@src/modules/account/domain';
 import { Group, GroupService, GroupTypes } from '@modules/group';
 import {
 	FederalStateService,
@@ -622,7 +622,7 @@ describe('OidcProvisioningService', () => {
 
 			it('should create a new account', async () => {
 				const { externalUser, schoolId, systemId, hash } = setupUser();
-				const account: AccountSaveDto = new AccountSaveDto({
+				const account: Account = new Account({
 					userId: 'userId',
 					username: hash,
 					systemId,
